+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 ## Improvements
 
 - PIM-9619: Improve error message when creating a new project with a name already used
+- RAC-509: Upgrade asset limit by asset family to 10 millions
 
 ## New features
 
@@ -22,3 +23,7 @@
 - Change parameter of `UnknownUserExceptionListener::onKernelException()` method from `Symfony\Component\HttpKernel\Event\GetResponseForExceptionEvent` to ` Symfony\Component\HttpKernel\Event\ExceptionEvent`
 - Change constructor of `Akeneo\Pim\WorkOrganization\Workflow\Bundle\Controller\ProductDraftController` to remove `Symfony\Component\Translation\TranslatorInterface $translator` parameter
 - Change constructor of `Akeneo\Pim\Automation\RuleEngine\Component\Engine\ProductRuleSelector` to remove `ProductRepositoryInterface $repo`
+- Remove class `Akeneo/AssetManager/back/Domain/Event/AssetFamilyAssetsDeletedEvent`
+- Update `Akeneo/AssetManager/back/Domain/Repository/AssetIndexerInterface.php` to:
+    - remove the `removeByAssetFamilyIdentifier` method
+